Model Parameters for a Log Normal Model

Model Parameters for a Log Normal Model

Usage

LogNormalModelParams(
  modname,
  nstates = 2,
  nmetro = 1000,
  forward = TRUE,
  cheat = FALSE,
  Insitu = InsituParams(),
  SurveilenceTest = SurveillanceTestParams(),
  ClinicalTest = ClinicalTestParams(),
  OutOfUnitInfection = OutOfUnitInfectionParams(),
  InUnit = LogNormalInUnitAcquisition(),
  Abx = AbxParams(),
  AbxRate = AbxRateParams()
)

LinearAbxModel(..., InUnit = LogNormalABXInUnitParameters())

Arguments

modname

The name of the model used. Usually specified by specification functions.

nstates

The number of states in the model.

nmetro

TODO

forward

TODO

cheat

TODO

Insitu

In Situ Parameters

SurveilenceTest

Surveillance Testing Parameters

ClinicalTest

Clinical Testing Parameters

OutOfUnitInfection

Out of Unit Infection Parameters

InUnit

In Unit Parameters, should be a list of lists with parameters for the acquisition, progression and clearance of the disease.

Abx

Antibiotic Parameters

AbxRate

Antibiotic Rate Parameters

Functions

LinearAbxModel(): Linear Antibiotic Model Alias